+++ src/detect-engine-file.c
@@ -36,7 +36,6 @@
 #include "detect-engine-hrhd.h"
 #include "detect-engine-hmd.h"
 #include "detect-engine-hcd.h"
-#include "detect-engine-hrud.h"
 #include "detect-engine-dcepayload.h"
 #include "detect-engine-file.h"
